+++ packages/terra-application-name/src/terra-dev-site/test/application-name/ApplicationMenuNameDefault.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import classNames from 'classnames/bind';
 import Image from 'terra-image';
 import ApplicationMenuName from '../../../ApplicationMenuName';
-import demoStyles from './demoStyles.scss';
+import demoStyles from './demoStyles.module.scss';
 
 const cx = classNames.bind(demoStyles);