+++ packages/terra-menu/src/terra-dev-site/test/menu/menu-item/MenuItemWrappedText.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import Menu from '../../../../Menu';
-import styles from './MenuItemCommon.scss';
+import styles from './MenuItemCommon.module.scss';
 
 const cx = classNames.bind(styles);