+++ packages/terra-navigation-layout/src/terra-dev-site/doc/example/MenuExample.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import classNames from 'classnames/bind';
 import PropTypes from 'prop-types';
 import { Link, withRouter } from 'react-router-dom';
-import styles from './NavigationLayoutDocCommon.scss';
+import styles from './NavigationLayoutDocCommon.module.scss';
 
 const cx = classNames.bind(styles);